+++ lib/webpay_rails/transaction_base.rb
@@ -1,6 +1,7 @@
 module WebpayRails
   class TransactionBase
-    def initialize(document)
+    def initialize(response)
+      document = Nokogiri::HTML(response.to_s)
       self.class.attr_list.each do |k|
         v = document.at_xpath("//#{k.to_s.tr('_', '')}")
         send("#{k}=", v.text.to_s) unless v.nil?